ouvy_saas/requirements.txt
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.0
Authlib==1.6.6
bandit==1.9.3
bleach==6.1.0
certifi==2026.1.4
cffi==2.0.0
cfgv==3.5.0
charset-normalizer==3.4.4
click==8.3.1
cloudinary==1.41.0
coverage==7.13.1
cryptography==46.0.3
distlib==0.4.0
dj-database-url==3.1.0
Django==6.0.1
django-cloudinary-storage==0.3.0
django-cors-headers==4.6.0
django-redis==6.0.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dparse==0.6.4
drf-yasg==1.21.12
filelock==3.20.3
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
identify==2.6.16
idna==3.11
inflection==0.5.1
iniconfig==2.3.0
Jinja2==3.1.6
joblib==1.5.3
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==4.2.0
mdurl==0.1.2
nltk==3.9.2
nodeenv==1.10.0
packaging==25.0
pillow==12.1.0
platformdirs==4.5.1
pluggy==1.6.0
pre_commit==4.5.1
psycopg2-binary==2.9.11
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.10.1
pytest==9.0.2
pytest-cov==7.0.0
pytest-django==4.11.1
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
qrcode==8.2
redis==7.1.0
regex==2026.1.15
requests==2.32.5
rich==14.2.0
ruamel.yaml==0.19.1
safety==3.7.0
safety-schemas==0.0.16
sentry-sdk==2.20.0
shellingham==1.5.4
six==1.17.0
sqlparse==0.5.5
stevedore==5.6.0
stripe==14.2.0
tenacity==9.1.2
tomlkit==0.14.0
tqdm==4.67.1
typer==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
virtualenv==20.36.1
webencodings==0.5.1
whitenoise==6.11.0
